Build spec modules against locally generated node headers tarball
